quantity/math: Add clarifying example about unit in sum/3
There is a general problem when summing over an empty list.
The reason stems from that the unit is unknown if the list is empty.

sum/3 will therefore let you specify a precision and unit so that
sum/3 can return a "zero" with a unit.

This commit adds examples about what is the expected result if there
is a mismatch between the type in the list and the fallback unit.